fix(pr232): move claude hooks into claude tab
